Extract login logic into services
