+++ app/src/services/localLogin.ts
@@ -1,0 +1,52 @@
+import axios from "axios";
+import getInvalidPasswordMessage from "../utils/getInvalidPasswordMessage";
+import getInvalidEmailMessage from "../utils/getInvalidEmailMessage";
+
+interface ILocalLoginProps {
+  email: string;
+  password: string;
+  setDone: React.Dispatch<React.SetStateAction<boolean>>;
+  setPasswordErrorMessage: React.Dispatch<React.SetStateAction<string>>;
+  setEmailErrorMessage: React.Dispatch<React.SetStateAction<string>>;
+}
+
+const localLogin = async ({
+  email,
+  password,
+  setDone,
+  setPasswordErrorMessage,
+  setEmailErrorMessage,
+}: ILocalLoginProps) => {
+  const errors = {
+    email: getInvalidEmailMessage(email),
+    password: getInvalidPasswordMessage(password),
+  };
+  setEmailErrorMessage(errors.email);
+  setPasswordErrorMessage(errors.password);
+  const thereAreErrors = !Object.values(errors).some((error) => error);
+  if (thereAreErrors) {
+    try {
+      const res = await axios({
+        method: "post",
+        url: "/auth/local",
+        data: {
+          email,
+          password,
+        },
+      });
+      if (res.status === 200) {
+        setDone(true);
+      }
+    } catch (error) {
+      if (error.response.status === 401) {
+        setPasswordErrorMessage("Invalid login!");
+        setEmailErrorMessage(" ");
+      } else {
+        setPasswordErrorMessage("Woppps something went wrong!");
+        console.error(error);
+      }
+    }
+  }
+};
+
+export default localLogin;
